Bump version to 2026.1-RC9 and update add method to accept double coordinates

+++ lib/src/main/java/io/github/udlepsprog2/prog2lib/program/GraphicsProgram.java
@@ -148,7 +148,7 @@ public abstract class GraphicsProgram extends Program {
      * @param y the y-coordinate of the object's location
      * @throws NullPointerException if {@code gObject} is {@code null}
      */
-    public void add(GObject gObject, int x, int y) {
+    public void add(GObject gObject, double x, double y) {
         canvas.add(gObject, x, y);
     }
